Console invokes the ramlReader on new raml data.
While the ramlReader is defined by the console and its output is
consumed by the console, it is invoked by the editor before
notifying the console of a change to the current raml file.

This commit inverts that relationship making the ramlReader a self
contained concern.

Additionally removed helpers.js which appears to be a duplication of the
ramlReader service.