The file browser now sports a context menu.
Hovering over a file in the file browser will show a gear icon. Clicking
on the gear will bring up a context menu. Save is currently the only
supported action from the context menu. The context menu can be brought
up for any file in the file browser, however it will not change focus to
that file if it is not currently being edited.